requirements_improved.txt
torch
numpy
pydub
transformers
soundfile
TTS
openai-whisper
PyQt5